Implement explicit sharded arithmetic operations using duck typing. (#31)
This level of math is aimed at the inner-most level of the device
hierarchy, where it is assumed that the devices connected to a process
are connected via high-speed links. By coding the model with explicit
sharded arithmetic, we naturally get multi-device distribution by having
a few points where we insert communication primitives. In this approach,
the same math expressions run on either unary or nary-sharded
tensor-parallel weight partitions. Using duck typing, the nary-sharded
partitions are seen as a `ShardedTensor` and corresponding duplication
of the math is undertaken. Unlike fully automatic solutions, this
approach *requires* that the model author does some minimal coding for
placement and distribution. In practice, this is unavoidable anyway, so
by taking the layers of automation out of it, it is at least
transparent, can be inspected eagerly, etc.

New ops:

* `elementwise`: Apply a sharded or unsharded unary/binary elementwise
operator.
* `sharded_cat`: Concatenate across the sharding dimension or return an
unsharded tensor as is.
* `sharded_sum`: Sum all shards or return unsharded tensor as is.

Adds implementation variants for:

* `matmul_sharded_rhs`: Matmul operating on a global lhs and a sharded
rhs. This is typically the first layer in a multi step sequence. Returns
a sharded tensor.
* `matmul_sharded`: Matmul operating on a sharded lhs and rhs. Performs
partial matmuls on each shard, returning a sharded result tensor.

It is left to the user, based on the arrangement of the math to choose
how to reduce a final result from a ShardedTensor to a unary one.

For now, not implemented on QuantizedTensors because additional infra
will be needed for row-wise sharding.