[shortfin] Plumb through await handling to Python asyncio. (#128)
* Stubs are in place that can be adapted for C++ coroutines but left for
later.
* Makes `Process` awaitable (for termination).
* Makes `ScopedDevice` awaitable (for sync).
* There's substantial room to profile/optimize this on the Python side,
but focusing on getting everything to connect first.

Handful of other changes:

* Adds CTest tests for some of the native support library.
* Adds a thread local to track the current native worker (saves a lot of
Python jumping).
* Introduces a CompletionEvent C++ class and associated Python binding.
This supports basic signaling and will be enhanced later with value
producing async objects.
* Namespaces the `iree_` support stuff under `iree::`
* Adds a BlockingExecutor to the system which can run arbitrary blocking
thunks. Uses this to workaround an issue with exporting hal semaphores
to wait_sources (this will get attention this week to remove the hack).
* Fixes a bad copy vs ref of slim_mutex, rendering mutexes all unique
(caught by unit tests). Deleted copy constructors from all classes not
meant to be copyable.
* Added a shared_event class which heap allocated and reference counts
an event. This is useful for various situations where an event producer
and consumer are separated in terms of object ownership.
* Added a device_sync example which demonstrates interleaved
fills/syncs.