[libshortfin] Fix illegal nested lock access in Message destruction. (#192)
* When Message instances had lifetime management transferred to Python,
destruction of the mutex would happen while it was held, causing an
assert. I believe this was only happening in some process level GC
scenarios before and was being masked. But it is also possible that it
is the cause of our spurious crashes.
* While this fix made sense once I found it, I also took the change to
get clang `-Wthread-safety` analysis working and wired up to a few of
the trickiest pieces of thread safety code.
* This revealed real bugs and design issues that are now fixed.