[libshortfin] Fix python stub generation. (#195)
This brings pylance back to life for the native bindings. It was giving
confusing error squiggles about not being able to use types that were
variables, but the root cause was that the backing stubs were not
generated properly for all of the module.

Includes a couple of other typing tweaks that were found to be helpful.