Quark dataset importer for fp8 (#96)
:edited:

Provides a entry point for quark models that are stored as a safetensor
file + config.json. With a little effort this could also be adapted to
become the equivalent of the hf importer from gguf without the
intermediate step.

I removed a lot of the useful debugging tooling I developed for this
because it doesn't play nicely with torch.export, but I imagine a lot of
it could be re-added if I guarded it in some way, so I'm keeping a
branch with those changes for reference.