Add reporting for llama testing and improve error handling (#322)
This commit generates reports for the llama testing. Every time it runs
it in CI, it will deploy to github pages
(https://nod-ai.github.io/SHARK-Platform). @aviator19941 is working on
getting the rest of the tests properly implemented. Only the llama8b
fp16 tests are working as intended at the moment.