Fp8 llama3 support (#320)
Makes changes to the importer to allow for newer quark format
Adds flag `--no-fake-quant` to enable exporting to mlir in native fp8
Adds support for KV cache quant/dequant
Adds support to QuantizerTensor to allow dequanting from a torch tensor
using previous parameters. This may not be the best way to do this, but
it seems better than pulling the quant parameters out into the wild.
This kind of exposes a missing abstraction in our workflow as most
things assume we are not using unpack().qs with the intention of doing
something with that value and then later dequantizing it. (The kv cache
being the relevant example)