SGLang Integration + Accuracy Tests, Restructure app_tests/integration_tests (#570)
# Description

This PR implements integration tests for the Shortfin LLM Server w/ the
SGLang integration. It uses llama3-8b-instruct on GPU, which is
downloaded using sharktank's `hf_datasets` script.

The tests server two purposes:
1. Test that the SGLang integration works properly at a functional
level.
2. Test that the accuracy of the responses from the shortfin LLM server
are consistent.
    - We have a batch of candidate questions, with expected answers
- We have temperature set to `1.0`, so the responses should be
deterministic.
 
This test is intended to run every 4 hours, which allows for us to
detect degradations in shortfin LLM output accuracy. If we do get a
failure due to an accuracy degradation, there will only be a small set
of shark-ai/iree commits that could be responsible.