In paged LLM exporting do not add any argument device affinities when not sharding (#576)
It is illegal to have ops with some arguments having affinties and some
not. Operands will be considered on different devices and this is not
allowed right now.
With recent enough version of IREE during compilation this will result
in

```
failure.mlir:1:1: error: affinity analysis failed to converge (input program may have invalid affinities assigned); use`--iree-stream-annotate-input-affinities` to help identify the invalid affinities
```

This change makes it so that we do not add any affinities when not
sharding.