Add T5 encoder exporting to MLIR and numerics verification with IREE (#573)
This change adds the export functionality and a test for the T5 v1.1
encoder. Small and XXL variants.
The numerical tolerance in the test is `alto=1e-4 rtol=2.0e-3`.
I observed relative error of `1.47e-3` on the XXL variant. This is
probably OK. Definitely not outrageous.
    
Added parsing of GGUF arrays when loading properties. This is used to
deduce the vocabulary size instead of having to provide this manually.
This can potentially be used also to load the tokenizer from GGUF or
IRPA.