See if stale dependencies are causing shortfin server to fail to start (#604)
# Description

We started seeing a failure in `Shortfin CPU LLM Integration Test` after
merging #601. However, the only aspect of the integration test that that
PR touches is a fix in the logger:

Old
```python
logger = logging.getLogger("__name__")
```

New
```python
logger = logging.getLogger(__name__)
```

That shouldn't have an impact on the test, and while reading the output
of the workflow, it didn't seem to be the line that caused the server to
not start.

When testing locally in a fresh environment, the test ran fine, which
made me think that it may be related to stale dependencies. I updated
the hash of cached pip to take into account requirement changes in
`sharktank` and `shortfin`, which appears to fix the test.