Add CLIP text model (#643)
Ports the CLIP text model from Hugging Face. This is the first iteration
so not much is changed from the original model. Things like dropout and
checkpointing are removed.
Add numeric verification tests for the various components of the stack
when executing in eager mode. Verifications are made for float32 and
bfloat16. There are tests for toy-sized components and the whole model
as well as the Large pretrained variant.
These tests does not include testing with IREE.

Functionalities for mask creation are not yet ported.