[tuner] Allow tuning of pipeline options (#791)
This PR adds some command line arguments to tune pipeline options. The
new flags are `--prefetch-shared-memory-options`,
`--no-reduce-shared-memory-bank-conflicts-options`, and
`--waves-per-eu-options`. The flags take a comma separated list of
values (bool or int), representing the possible values that the
corresponding pipeline options can take in tuning.

This PR also adds `promote_operands = [0, 1]` to the TileAndFuse tuning
configurations. This could potentially be a tuned parameter in the
future, but for now it is typically good to promote the lhs and rhs
operands.

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>