Bump IREE version pins to 3.2.0rc20250109. (#802)
See https://github.com/nod-ai/shark-ai/issues/760 for context. We want
to stay close to the latest versions while still pinning versions for
predictability. Updating version pins is currently a manual process but
we plan on automating it in the future.

We can decide how noisy we want these dependency updates to be:

* new PRs daily or less frequently
* do or don't reuse existing PRs
* merge ASAP or let them sit for multiple days