Make path manipulation play nice with symlinks by deferring symlink resolution (#749)
pathlib.Path.resolve() would follow and resolve symlinks, which is
usually not what we want. In most cases we want to defer symlink
resolution.
When doing path manipulation we may care about the surrounding
directory structure, which may be a collection of multiple symlinks, so
doing a symlink resolution would put you in a different directory
structure, which is probably not the intent.

This change makes all path manipulations of __file__ not resolve first.
Since Python 3.9 __file__ is absolute
https://docs.python.org/3/whatsnew/3.9.html#other-language-changes
It is not stated explicitly that it should normalized as well but we
assume it.
We use os.path.abspath if required to get unresolved, absolute and
normalized path.