[Llama] Remove inplace read for KVCache (#849)
The in-place read for KVCache doesn't seem to be used and is just used
to tell the KVCache number of partitions to read from. The number of
partitions for KVCache is in reality fixed (as from the name, K/V cache)
and we are adding overhead by passing this parameter.