[Sharktank][Llama][FP8] Minimal changes for numerically correct fp8 (#859)
This patch enables the use of quark quantized models of the latest
generation. Many changes were required to enable parity with the source
model which is very sensitive to any numerical fluctuations. There is a
test added to maintain this parity but is disabled until I can get the
relevant data set up on the ci machine.