Move server-specific config args out of ModelParams and into new ServerParams in config_struct.py (#877)
# Problem

Currently, prefix sharing is read as a config key from the
exported-model's config.json.

There is no reason for it to be there - - this is not a property of the
model, but rather a property of the server instance that's being
launched.

This overcomplicates the tests: rather than just inserting an additional
cmdline arg, the tests have to make an edited copy of the exported
config files in order to specify the prefix-sharing algorithm.

# Solution
This PR puts in a ServerParam class for storing config options that
don't require re-exporting the model. Right now it just affects the
prefix-sharing algorithm but this makes space for things like:
- decoding / token search options
- multi-token prediction

# Why this?

Deepseek v3 will have many run-time options that aren't relevant at
export/compile, e.g.
[MTP](https://huggingface.co/facebook/multi-token-prediction). These
changes will make it easier to set those up.