Add `log_softmax` op to `shortfin.array` (#1076)
# Desciption

This adds `log_softmax` to the `shortfin.array` API, which is a required
function for `beam_search`, `sampling`, and any other selection method
that relies on cumulative probability.

`numpy` doesn't have a `log_softmax` function out of the box, which is
what `xtensor` typically [maps
to](https://xtensor.readthedocs.io/en/latest/numpy.html).

So, I wrote it to ensure that it matches
the same behavior as `torch.log_softmax`