Rework batcher for a reservation system (#1183)
With decode the majority of future steps are known a groupable. Created
a basic reservation system so that steps can be reserved and the
continuous batcher knows that a future groupable request is coming.

This should be adapted to include grouping multiple responses, and
having a "grouping" exist. E.g. multiple parallel requests in flight and
grouped together for the multiple hypothesis case.