[sharktank] Align the meaning of ops.equal to that of torch.equal (#1227)
The previous meaning was not precise and diverged from that of torch:
Compares 2 tensors for equality, such that if one is substituted with
the other in sharktank polymorphic calls, the results will be
essentially the same. Meaning, they would also compare equal.

We should try to not change the meaning of ops compared to torch. Now
the meaning is the same as in torch. They are equal if their dtypes is
equal and their elements.
For that we utilize unboxing.

Add parameter to is_deep_equal to not require name equality.