Shortfin llm `top_k` + `top_p` sampling (#1259)
Stacked on #1234 

# Description

This PR implements the
[top_p](https://huggingface.co/blog/how-to-generate#top-p-nucleus-sampling)
(or `nucleas`) sampling strategy for the Shortfin LLM server. It also
allows the use of `top_k + top_p` in series with each other, which I
found to be supported by most LLM servers.

## Explanation

`top_p` is a sampling algorithm, where per-beam we sort the logits of a
decode invocation. We then select tokens from those logits, until the
cumulative probability of those tokens are equal to `top_p`.

We then randomly sample from that distribution by weight, where the
number of samples we take depends on the `token_selection_strategy`.

If `top_k` and `top_p` are provided, we first do `top_k` sampling, then
`top_p` sampling on the results from `top_k`. The idea is that you can
control the subset of tokens that `top_p` is going to sample from. So, a
low `top_k` can restrict the subset of tokens to select from, making
`top_p` still diverse, but less random overall. A high `top_k` value can
increase the size of the subset that `top_p` samples from, and
potentially introduce more variation, depending on the tokens that
`top_k` randomly selects.

This also includes some simplifications on the overall interface,
allowing beams to share the sample `sample_top_k` and `sample_top_p`
functions.

---------

Co-authored-by: Rob Suderman <rob.suderman@gmail.com>