[tuner] padding support along TileAndFuse pipeline (#1310)
Motivated by bug reported from this
[input](https://gist.github.com/rkayaith/0c537b5b917488a890273f7d886791c7)

**Problem:**
In the current tuner, tile sizes are required to both evenly divide the
problem size and be multiples of the intrinsic size. When the problem
size is not a multiple of the intrinsic size, the solver fails to find
any valid solutions.

**Solution:**
The simplest fix is to round the problem size up to the nearest multiple
of the intrinsic size when generating constraints, and then account for
the necessary padding in the lowering configuration.

This PR introduces two main changes from Max's local experimental tuning
branch into the main tuner branch:

- Change 1: modify the constraint generation to support unaligned
problem shapes.
- Change 2: add padding information to the lowering configuration.
- Include corresponding tests for the above changes.

---------

Signed-off-by: Bangtian Liu <liubangtian@gmail.com>