Refactor pipeline parallelism work and fix block offset (#1379)
Pipeline parallelism had issues with transformer block offset. Updated
and cleaned up the block / device mappings to instead use two separate
mappings. This allows a cleaner implementation that relies less on
replicated behavior.

Included an `unshard` operation for the KV cache as the existing tests
replicated the reconstruction which should be part of PagedAttention.

---------

Co-authored-by: Alex Vasile <48962821+Alex-Vasile@users.noreply.github.com>