[sharktank] Add exporting of sharded toy sized Resnet block IREE test data (#1033)
Export test data to add a test in the IREE test suite.

Refactor the test in sharktank to utilize the shared export function.
Make the test not xfail due to numerical precision as the error is
around 1e-5 for f32 which should be good enough.

2 other tests required adjusting the numerical tolerance due to a change
in random theta generation such that it does not generate in range [0,
1], but in range [-1, 1].