[RoPE] Cleanup RoPE implementation (#1514)
Also deletes the attention_test.py test which was simply wrong and there
is no way to fix it. The test passed hidden states of the sharktank
implementation to the hugging face llama model which can obviously not
be always consistent. The test also passed a wrong dictionary as model
config to the hugging face implementation which is also wrong, because
we could never pass the llama rope scaling to it.