[sharktank] Fix which prefill logits IREE perplexity calculations are using (#1535)
The current implementation had two issues:
1. IREE and eager perplexity calculations were using different prefill
logits
2. Neither were taking into account the sequence length of a given
prompt within the batch when grabbing a slice of the logits.

This PR changes the IREE perplexity calculation to use the correct
prefill logits; each entry in the batch the slice corresponding to the
final real token is being used to calculate perplexity.

Only fixing the IREE side, the eager side is untouched (and left
incorrect) because of #1509 prevent rerunning to get new values. Leaving
eager calc as is also allows for #1509 to be more easily fixed.

TODO:
- [x] Update perplexity scores in `baseline_perplexity_scores.json`

For: #1495

---------

Signed-off-by: Alex Vasile <48962821+Alex-Vasile@users.noreply.github.com>
Co-authored-by: Archana Ramalingam <98564406+archana-ramalingam@users.noreply.github.com>