Reland caching allocator for avoiding reallocations (#1519) (#1561)
We can avoid reallocating at the hip level by cacheing allocations
between invocations. This is particularly useful for decode which sees
sets of 32 decode steps with the same dimensions and dtypes.

Previous version failed to remove reused buffers from the list of cached
tensors. This resulted in potential buffer reuse and could cause a crash
when attempting to find that cached buffer in the future.

Caching limit was also changed to rebuild the cache / tables rather than
mutate them. This avoids unintended effects of the `del` operator.