[sharktank] do not use scatter_add in MoE to improve performance (#1570)
Since there is regularity in the index we just do a reduction instead of
scatter_add.

This is a fix for IREE performance issue
https://github.com/iree-org/iree/issues/20968 and numerical bug in IREE
with scatter_add discovered by @Alex-Vasile.

Increased the IREE perplexity test delta compared to baseline as it was
off buy more than expected. This test is anyway on the toy model so not
much meaning can be put into this number.