Change perplexity logit padding to align logits with tokens (#1588)
Logits need to be padded to keep them the same length as token_ids since
we don't start with token_ids[:, 0], we instead prefil will a range of
tokens to give better context to the model.

Previously the logits were right padded based on the difference between
in length between their size and that of token_ids. This causes
token_ids[:, i] and logits[:, i] to no longer line up requiring special
indexing further down.

This PR cleans up that padding approach to make the indexing consistent.

---------

Signed-off-by: Alex Vasile <48962821+Alex-Vasile@users.noreply.github.com>