Add `write_range` function to `KVCache` (#1724)
This is part of the preparation work for enabling `partial_prefill`
(prefill with `start_positions` >= 0).

It's similar to `write_timestep`, except extends the functionality to be
able to write to kvcache for `seq_len` > 1. During partial_prefill, we
will use this function to write the `k` and `v` tensors corresponding to
`sequence(i)[start_position(i):]` for each sequence in a batch.

Afterwards, we'll read from the cache, avoiding a recomputation of
already cache `k` and `v` tensors corresponding to
`sequence(i)[:start_position(i)]`