[Dataset] Add json load/save support to Dataset (#2070)
This patch adds json support to save / load dataset as empty tensors.
This is designed to allow exporting models without needing the whole
irpa file for repros.

For example, for llama-405b, the entire irpa file is 207GB, but the json
file is 500kb (unseralized with indenting).

Main changes:

- Split add_to_archive method on InferenceTensor to get_metadata and
add_to_archive, preserving add_to_archive behavior.
- Changes add_to_archive behavior for ReplicatedTensor to save the
tensor name as the first shard, this is more consistent with how the raw
entries are stored during runtime, and its better to save them in the
same way.
- Adds cli dataset support for --json-file = 
- Add a tool to dump a dataset (gguf/irpa/json) to json