[LLM Server] Enable Chunked Prefill (#2221)
Stacked on #2208, please review that PR first, as I'll pull the changes
from there into here.

We should be able to merge while we wait on #2235 

## Description

This PR enables `ChunkedPrefill` for the shortfin server. It creates a
simple `ChunkScheduler`, where really the only difference is that it
references a `_pending` map of `rid => List[LlmTaskInput]`. When it's
notified that a chunk has finished, the scheduler moves the next chunk
from that list to the `_ready` queue. From there, it pretty much works
the same as the existing `Scheduler`.

It also adds the function to our `default batcher` that allows us to
split an `LlmInferenceExecRequest` into a `List[LlmTaskInput]`, which we
then submit to the scheduler.