Llm fix unintended chunking (#2280)
Since I was defaulting the server cli value to `8`, we were chunking
when we weren't wanting to.

However, this and #2221 does not seem to be the cause of #2279. 

I was able to download the dataset being used in the mlperf harness, run
a local shortfin 8b server, and checkout #2250 locally (one before), and
am seeing the same reported bug:

```bash
:0:rocdevice.cpp            :2991: 123228313452 us:  Callback: Queue 0x7fa575000000 aborting with error : HSA_STATUS_ERROR_MEMORY_APERTURE_VIOLATION: The agent attempted to access memory beyond the largest legal address. code: 0x29
```