[sharktank] use our get_logger instead of logging.getLogger (#2277)
This allows us to control the log level with env var TURBINE_LOG_LEVEL.

E.g.
```
export TURBINE_LOG_LEVEL=INFO
```

CI tests are not augmented to with this env var setting as it is
expected that https://github.com/iree-org/iree-turbine/pull/1155 will
get merged to allow pytest to also control the log level.

In the future we could potentially refactor our `get_logger` to control
with a different env var.