[Fusilli] Dump scheduling info json and check for number of dispatches (#2359)
Tracks an additional compilation statistics json as compiled & cached
asset. This is useful to ensure we're tracking regressions to the number
of dispatches (and any other useful stats in the future) in the
generated compiled binary in CI.

Sets `FUSILLI_CACHE_DIR` to `/tmp` for testing only - this was required
to have the lit driver let the test executable populate and read from
the json on disk.

Updates LIT tests for most conv cases (different memory layouts,
pad/nopad) and takes a TODO to update dispatch count for the pad cases
once @IanWood1 's IREE fixes in
https://github.com/iree-org/iree/issues/22091 land and are picked in
Fusilli.

For now the assets are cleaned up when `remove=True`, however in a
follow-on we'll want to make the dumping of the json also opt-in (gated
by a debug/testing-only flag).