[sharktank] Add tests and fixes for pipeline-parallel Llama 405B and Toy f4 vs non-parallel (#2353)
Add tests to compare perplexity of PP vs non-PP with a small set of
random tokens.

The Llama 405B f4 test is for a pruned (removed layers) model as we
don't care if the perplexity is high. Just that it is the same against
the reference.

Fix LlmPerplexityEval such that it frees the pages that it allocates.

Make `ops.attention_mask` trivially replicable.

Make topk count in decode for the LLM utils a configurable argument
instead of always hardcoding to 8.