[sharktank] Llama 3.1 f16 HF import presets (#2290)
Add registry for import presets and populate it with Llama 3.1 f16
models.

Add support for referencing a HF dataset during import. This decouples
specification of what needs to be downloaded versus how to import the
dataset after it is download.

Expand the HF datasets to specify the model files not completely
explicitly, but by using filters like
`huggingface_hub.snapshot_download`.

Next steps are:
1. Make the CI use this new mechanism.
2. Add importation of models with more complicated transformations like
quantization.