[Sharktank] Scheduler for `vmfb-runner` (#2427)
# Description

- Introduce very simple `Scheduler` component
- Split components into separate files
- Modify `LlmDecoder` to use new flow

Right now, just focusing on `LLmDecoder`, but should be able to enable
things like the ppl tests with minimal changes. Namely, making the
`selection_fn` optional for `run_prefill` and `run_decode` and just
returning the logits at the end. For now, I left the existing `prefill`
and `decode` functions to minimize stages, and so PPL can be enabled
with this flow as its own issue.

This also enables us to run any size dataset. Previously, we would throw
in error if, for example, `bs_prefill == 4` and we tried to run with `5
prompts`.

With these changes, you can request any number of prompts, and it works
correctly.