[sharktank] improve module patching for tensor tracing (#2239)
Make patching be able to patch/wrap not just the forward method but
select what methods to patch based on a regex or fnmatch-like match on
the fully qualified name of submodule methods.

When using TraceTensorModulePatch always trace with a key like
SaveModuleResultTensorsPatch. This makes both patchers consistent.
We probably  should retire `SaveModuleResultTensorsPatch` at some point.

---------

Co-authored-by: Archana Ramalingam <98564406+archana-ramalingam@users.noreply.github.com>