[Fusilli] Add rank checks to convFPropNode validation (#2524)
When adding a test for conv3d, I was hitting a strange crash without
anything useful to go by. After some debugging, realized that the
padding shapes did not match what conv3d expected. This should ideally
be caught in the graph validation phase, so this PR adds checks for:
- [x] inputs, weights and outputs have matching rank
- [x] padding, stride and dilation match the number of spatial dims