[Fusilli] Check for consistency in ConvWGradNode's input/output shapes (#2585)
Refactors conv output shape inference logic into a helper function
`getConvInferredOutputShape` so that it can be used by `ConvWGradNode`
to verify that the Y/DX/DW tensors have consistent shapes.


This is motivated by a strange error I was hitting in `iree-compile`
that ended up being because I specified the wrong shape for `DW`. This
change should make the source of the problem more clear.

---------

Signed-off-by: Ian Wood <ianwood@u.northwestern.edu>