[Fusilli] Add support for ConvWGrad and ConvDGrad to benchmark driver (#2624)
This PR adds support to `benchmark/driver.cpp`, specifically the
`benchmarkConvWGrad` and `benchmarkConvDGrad` functions that are used in
the CLI, combined with the `-F` flag to specify which to run (1, 2, 4
and for forward, weight, and data respectively. This is based off the
boo driver's args, but they can be changed. I don't think 1, 2, 4 is
particularly intuitive). Updated and added tests to reflect this.

Note: currently, Conv3D for WGrad and DGrad seems to be broken. See this
issue: https://github.com/nod-ai/shark-ai/issues/2619 for more details.
This is why they are not present in the tests.

---------

Signed-off-by: raayandhar <rdhar@amd.com>