[sharktank] Materialize indices in PrefillTask to ease branching logic in downstream code (#2507)
Handling the the None special case is error prone and adds branching
complexity. We don't care that much about performance when doing
evaluation.

Allow compute_cross_entropy to accept sequences of logits and indicies
instead of single ndarray. This makes the function symmetric with
respect to the requests arg. Which is a sequence.
This will help with chunked prefill and other methods of constructing
the results where they are stitched from multiple requests. We don't
need to burden the caller to make a single ndarray when we don't need
it. But a single ndarray is still supported.